@@ -1,6 +1,6 @@
 {
   "name": "tv2-sofie-blueprints-inews",
-  "version": "1.7.0",
+  "version": "1.7.1",
   "repository": "https://github.com/olzzon/tv2-sofie-blueprints-inews",
   "license": "MIT",
   "private": true,
@@ -35,31 +35,34 @@
     ]
   },
   "devDependencies": {
-    "@types/jest": "^24.0.0",
+    "@types/jest": "^27.4.1",
     "@types/node": "^12.12.2",
     "@types/underscore": "^1.8.9",
     "axios": "^0.19.0",
     "git-revision-webpack-plugin": "^3.0.3",
-    "jest": "^24.0.0",
+    "jest": "^27.5.1",
     "jest-haste-map": "^24.5.0",
     "jest-resolve": "^24.5.0",
-    "moment": "^2.22.2",
+    "moment": "^2.29.2",
     "node-license-validator": "^1.3.0",
     "prettier": "^1.18.2",
     "standard-version": "9.1.1",
-    "ts-jest": "^24.0.0",
+    "ts-jest": "^27.1.3",
     "ts-loader": "^6.2.1",
     "ts-node": "^8.0.3",
     "tsconfig-paths-webpack-plugin": "^3.2.0",
     "tslint": "^5.12.1",
     "tslint-config-prettier": "^1.18.0",
     "tslint-plugin-prettier": "^2.0.1",
-    "typescript": "^3.3.3",
+    "typescript": "^4.1",
     "webpack": "^4.27.1",
     "webpack-cli": "^3.1.2"
   },
   "dependencies": {
     "@tv2media/blueprints-integration": "v1.37.1-1",
     "underscore": "^1.12.1"
+  },
+  "resolutions": {
+    "moment": "^2.29.2"
   }
 }
