@@ -13,11 +13,11 @@ import { literal } from '../util'
 
 export * from './moveSourcesToTable'
 export * from './addKeepAudio'
-export * from './shortcuts'
 export * from './transitions'
 export * from './graphic-defaults'
 export * from './manifestWithMediaFlow'
 export * from './sourceManifest'
+export * from './sourceLayers'
 export * from './forceSourceLayerToDefaultsBase'
 export * from './hotkeys'
 
