@@ -90,7 +90,8 @@ export enum AdlibTags {
 	ADLIB_RECALL_LAST_LIVE = 'recall_last_live',
 	ADLIB_RECALL_LAST_DVE = 'recall_last_dve',
 	ADLIB_SELECT_DVE_LAYOUT = 'select_dve_layout',
-	ADLIB_TAKE_WITH_TRANSITION = 'take_with_transition'
+	ADLIB_TAKE_WITH_TRANSITION = 'take_with_transition',
+	ADLIB_FADE_DOWN_PERSISTED_AUDIO_LEVELS = 'fade_down_persisted_audio_levels'
 }
 
 /**
@@ -133,7 +134,9 @@ export enum AdlibActionType {
 	CLEAR_GRAPHICS = 'clear_graphics',
 	TAKE_WITH_TRANSITION = 'take_with_transition',
 	RECALL_LAST_LIVE = 'recall_last_live',
-	RECALL_LAST_DVE = 'recall_last_dve'
+	RECALL_LAST_DVE = 'recall_last_dve',
+	FADE_DOWN_PERSISTED_AUDIO_LEVELS = 'fade_down_persisted_audio_levels',
+	PLAY_GRAPHICS = 'play_graphics'
 }
 
 export enum TallyTags {
