@@ -10,7 +10,6 @@ function getSisyfosLayers(configName: string, id: string): string[] {
 			return []
 		case 'SourcesRM':
 		case 'SourcesFeed':
-		case 'SourcesSkype':
 			return ['sisyfos_source_live_' + id]
 		case 'SourcesDelayedPlayback':
 			return ['sisyfos_source_evs_' + id]
@@ -24,13 +23,13 @@ function prepareConfig(
 	conf: string,
 	configName: string,
 	studioMics: boolean,
-	keepAudioInStudio?: boolean
+	wantsToPersistAudio?: boolean
 ): Array<{
 	SourceName: string
 	AtemSource: number
 	SisyfosLayers: string[]
 	StudioMics: boolean
-	KeepAudioInStudio: boolean
+	wantsToPersistAudio: boolean
 }> {
 	return parseMapStr(undefined, conf, true).map(c => {
 		return {
@@ -38,7 +37,7 @@ function prepareConfig(
 			AtemSource: c.val,
 			SisyfosLayers: getSisyfosLayers(configName, c.id),
 			StudioMics: studioMics,
-			KeepAudioInStudio: keepAudioInStudio ?? false
+			wantsToPersistAudio: wantsToPersistAudio ?? false
 		}
 	})
 }
@@ -89,7 +88,6 @@ export const defaultStudioConfig: StudioConfig = {
 		true
 	),
 	// TODO: prepareConfig is legacy code, refactor when refactoring FindSourceInfo
-	SourcesSkype: prepareConfig('1:1,2:2,3:3,4:4,5:5,6:6,7:7', 'SourcesSkype', false),
 	SourcesRM: prepareConfig('1:1,2:2,3:3,4:4,5:5,6:6,7:7,8:8,9:9,10:10', 'SourcesRM', false, true),
 	SourcesFeed: prepareConfig('1:1,2:2,3:3,4:4,5:5,6:6,7:7,8:8,9:9,10:10', 'SourcesFeed', false, true),
 	SourcesDelayedPlayback: prepareConfig('1:5,2:5', 'SourcesDelayedPlayback', false),
