@@ -14,7 +14,8 @@ import {
 	CueDefinitionTelefon,
 	GraphicInternal,
 	literal,
-	PartDefinitionKam
+	PartDefinitionKam,
+	PieceMetaData
 } from 'tv2-common'
 import { CueType, PartType, SharedGraphicLLayer, SharedOutputLayers } from 'tv2-constants'
 import { SegmentUserContext } from '../../../../__mocks__/context'
@@ -90,8 +91,6 @@ describe('telefon', () => {
 				studio: (defaultStudioConfig as unknown) as StudioConfig,
 				sources: [],
 				mediaPlayers: [],
-				stickyLayers: [],
-				liveAudio: [],
 				dsk: defaultDSKConfig,
 				selectedGraphicsSetup: DEFAULT_GRAPHICS_SETUP
 			},
@@ -114,6 +113,11 @@ describe('telefon', () => {
 				outputLayerId: SharedOutputLayers.OVERLAY,
 				sourceLayerId: SourceLayer.PgmGraphicsTLF,
 				lifespan: PieceLifespan.WithinPart,
+				metaData: literal<PieceMetaData>({
+					sisyfosPersistMetaData: {
+						sisyfosLayers: []
+					}
+				}),
 				content: literal<WithTimeline<GraphicsContent>>({
 					fileName: 'bund',
 					path: 'bund',
