@@ -11,7 +11,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 10,
 		name: 'GFX Ident',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
@@ -29,7 +29,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 10,
 		name: 'GFX Ident Persistent (hidden)',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
@@ -47,7 +47,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 20,
 		name: 'GFX Top',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
@@ -65,7 +65,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 30,
 		name: 'GFX Lowerthirds',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
@@ -83,7 +83,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 40,
 		name: 'GFX Headline',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
