@@ -33,8 +33,7 @@ export const OFFTUBE_DVE_GENERATOR_OPTIONS: DVEOptions = {
 		},
 		SisyfosLLayer: {
 			ClipPending: OfftubeSisyfosLLayer.SisyfosSourceClipPending,
-			StudioMics: OfftubeSisyfosLLayer.SisyfosGroupStudioMics,
-			PersistedLevels: OfftubeSisyfosLLayer.SisyfosPersistedLevels
+			StudioMics: OfftubeSisyfosLLayer.SisyfosGroupStudioMics
 		},
 		CasparLLayer: {
 			ClipPending: OfftubeCasparLLayer.CasparPlayerClipPending
@@ -45,8 +44,7 @@ export const OFFTUBE_DVE_GENERATOR_OPTIONS: DVEOptions = {
 		GetLayersForEkstern
 	},
 	boxMappings,
-	AUDIO_LAYERS: [], // TODO
-	EXCLUDED_LAYERS: [] // TODO
+	AUDIO_LAYERS: [] // TODO
 }
 
 export function OfftubeMakeContentDVE(
@@ -57,7 +55,7 @@ export function OfftubeMakeContentDVE(
 	dveConfig: DVEConfigInput | undefined,
 	addClass?: boolean,
 	adlib?: boolean
-): { content: WithTimeline<SplitsContent>; valid: boolean; stickyLayers: string[] } {
+): { content: WithTimeline<SplitsContent>; valid: boolean } {
 	return MakeContentDVEBase(
 		context,
 		config,
