@@ -11,7 +11,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 10,
 		name: 'GFX Ident',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
@@ -26,7 +26,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 10,
 		name: 'GFX Ident Persistent (hidden)',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
@@ -41,7 +41,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 20,
 		name: 'GFX Top',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
@@ -56,7 +56,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 30,
 		name: 'GFX Lowerthirds',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
@@ -71,7 +71,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 40,
 		name: 'GFX Headline',
 		abbreviation: 'G',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
@@ -101,7 +101,7 @@ const OVERLAY: ISourceLayer[] = [
 		_rank: 60,
 		name: 'GFX Telefon',
 		abbreviation: 'TLF',
-		type: SourceLayerType.GRAPHICS,
+		type: SourceLayerType.LOWER_THIRD,
 		exclusiveGroup: '',
 		isRemoteInput: false,
 		isGuestInput: false,
