@@ -21,7 +21,6 @@ import {
 	manifestOfftubeDownstreamKeyers,
 	manifestOfftubeSourcesABMediaPlayers,
 	manifestOfftubeSourcesCam,
-	manifestOfftubeSourcesRM,
 	manifestOfftubeStudioMics
 } from '../config-manifests'
 import { OfftubeCasparLLayer, OfftubeSisyfosLLayer } from '../layers'
@@ -205,11 +204,29 @@ export const studioMigrations: MigrationStepStudio[] = literal<MigrationStepStud
 	ensureStudioConfig(
 		'0.1.0',
 		'SourcesRM',
-		manifestOfftubeSourcesRM.defaultVal,
+		[
+			{
+				_id: '',
+				SourceName: '1',
+				AtemSource: 3,
+				SisyfosLayers: [OfftubeSisyfosLLayer.SisyfosSourceLive_3],
+				StudioMics: true,
+				KeepAudioInStudio: true
+			}
+		],
 		'text',
 		'Studio config: Remote mappings',
 		'Enter the remote input mapping',
-		manifestOfftubeSourcesRM.defaultVal
+		[
+			{
+				_id: '',
+				SourceName: '1',
+				AtemSource: 3,
+				SisyfosLayers: [OfftubeSisyfosLLayer.SisyfosSourceLive_3],
+				StudioMics: true,
+				KeepAudioInStudio: true
+			}
+		]
 	),
 
 	ensureStudioConfig(
@@ -309,7 +326,16 @@ export const studioMigrations: MigrationStepStudio[] = literal<MigrationStepStud
 	 * - Split RM config into FEED and RM configs
 	 * - Add concept of roles to DSK config table (and cleanup configs replaced by table)
 	 */
-	SetConfigTo('1.6.1', 'Offtube', 'SourcesRM', manifestOfftubeSourcesRM.defaultVal),
+	SetConfigTo('1.6.1', 'Offtube', 'SourcesRM', [
+		{
+			_id: '',
+			SourceName: '1',
+			AtemSource: 3,
+			SisyfosLayers: [OfftubeSisyfosLLayer.SisyfosSourceLive_3],
+			StudioMics: true,
+			KeepAudioInStudio: true
+		}
+	]),
 	SetConfigTo('1.6.1', 'Offtube', 'AtemSource.DSK', manifestOfftubeDownstreamKeyers.defaultVal),
 	RemoveConfig('1.6.1', 'Offtube', 'AtemSource.JingleFill'),
 	RemoveConfig('1.6.1', 'Offtube', 'AtemSource.JingleKey'),
@@ -329,6 +355,8 @@ export const studioMigrations: MigrationStepStudio[] = literal<MigrationStepStud
 	 * - Force soundbed caspar layer to defaults (channel 3, layer 101)
 	 */
 	GetMappingDefaultMigrationStepForLayer('1.6.10', OfftubeCasparLLayer.CasparCGLYD, true),
+
+	RenameStudioConfig('1.6.11', 'Offtube', 'SourcesRM.KeepAudioInStudio', 'SourcesRM.WantsToPersistAudio'),
 
 	// Fill in any mappings that did not exist before
 	// Note: These should only be run as the very final step of all migrations. otherwise they will add items too early, and confuse old migrations
